Aggressively try to use Integer objects
When doing math, try to do integer arithmetic when the arguments passed
are Integer objects, rather than always falling back to floating-point
math. This speeds up some benchmarks that rely heavily on integer
arithmetic.

+++ rhino/src/main/java/org/mozilla/javascript/ArrayLikeAbstractOperations.java
@@ -120,7 +120,8 @@ public class ArrayLikeAbstractOperations {
                     break;
                 case FIND_INDEX:
                 case FIND_LAST_INDEX:
-                    if (ScriptRuntime.toBoolean(result)) return ScriptRuntime.wrapNumber(i);
+                    if (ScriptRuntime.toBoolean(result))
+                        return ScriptRuntime.wrapNumber((double) i);
                     break;
             }
         }
